+++ main_code/lib/Common/Common.cpp
@@ -1,16 +1,16 @@
 #include "Common.h"
 
-double rad2deg(double angle) { return angle * 180 / pi; }
+float rad2deg(float angle) { return angle * 180 / pi; }
 
-double deg2rad(double angle) { return angle * pi / 180; }
+float deg2rad(float angle) { return angle * pi / 180; }
 
 float angleDiff(float angle1, float angle2) {
     float diff = fmod(angle1 - angle2, 360);
     return min(abs(diff), 360 - abs(diff));
 }
 
-bool angleIsInside(double angleBoundCounterClockwise,
-                   double angleBoundClockwise, double angleCheck) {
+bool angleIsInside(float angleBoundCounterClockwise,
+                   float angleBoundClockwise, float angleCheck) {
     if (angleBoundCounterClockwise < angleBoundClockwise) {
         return (angleBoundCounterClockwise < angleCheck &&
                 angleCheck < angleBoundClockwise);
@@ -21,13 +21,13 @@ bool angleIsInside(double angleBoundCounterClockwise,
 }
 
 // normalise to 0 to 1
-double norm(float val, int max, int min) { return (val - min) / (max - min); }
+float norm(float val, int max, int min) { return (val - min) / (max - min); }
 
-double nonReflex(double angle) {
+float nonReflex(float angle) {
     if (angle > 180) return angle - 360;
     return angle;
 }
 
-double distance(double x, double y) { return sqrt(x * x + y * y); }
+float distance(float x, float y) { return sqrt(x * x + y * y); }
 
 int sign(int x) { return (x >= 0) ? 1 : -1; }
